% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Correlation coefficient}
\usage{
corr(xs, ys)
}
\arguments{
\item{xs}{a (non-empty) numeric matrix of data values}

\item{ys}{a (non-empty) numeric matrix of data values}
}
\value{
the correlation coefficient of multiple pairs of vectors
}
\description{
Simultaneous calculation of correlation coefficients of multiple pairs of vectors
}
\examples{
\dontrun{
x <- replicate(5, runif(10), simplify = FALSE)
y <- replicate(5, rnorm(10), simplify = FALSE)
corr(x,y)
}
}
