% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.stat.R
\name{cvm.stat}
\alias{cvm.stat}
\title{Cramer-von Mises statistic}
\usage{
cvm.stat(x, y)
}
\arguments{
\item{x}{a numeric vector of data values}

\item{y}{either a numeric vector of data values}
}
\value{
the numeric value of the statistic
}
\description{
calculate the two-sample Cramer-von Mises statistic value
}
\examples{
\dontrun{
x <- rnorm(50)
y <- runif(30)
# Do x and y come from the same distribution?
cvm.stat(x, y)

# test if x is stochastically larger than x2
x2 <- rnorm(50, -1)
plot(ecdf(x), xlim = range(c(x, x2)))
plot(ecdf(x2), add = TRUE, lty = "dashed")
cvm.stat(x, x2)
}
}
